backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
langchain==0.1.16
langchain-openai==0.0.8
